Adding new namelists to pwscf
Also adding helper function for transforming values to fortran
